update_角色系統

+++ 3D_GAME_20220319/Assets/Scripts/DateEnemy.cs
@@ -39,7 +39,7 @@ namespace Jones
         [Header("攻擊區域尺寸")]
         public Vector3 v3AttackSize=Vector3.one;
         [Header("經驗值"), Range(0, 1000)]
-        public float exp;
+        public int exp;
         [Header("金幣物件")]
         public GameObject goCoin;
         [Header("金幣數量"), Range(0, 1000)]
